Corrections to lagoon interaction commands documentation
In a previous pull request (https://github.com/uselagoon/lagoon-cli/pull/243) I noted that the command for changing the default lagoon required the use of the --lagoon flag in order to succeed and this change was merged in. 

I have gone back to verify the rest of the documentation in this section and have identified that this same requirement exists for the other commands including:

* Adding a lagoon
* Deleting a lagoon

Attempting to execute the examples given in the documentation without the usage of the --lagoon flag will cause the operations to fail with:

```
Error: Missing arguments: Lagoon name is not defined
```

My change updates both the command documentation and the practical examples to demonstrate the required inclusion of this flag when interacting with the Lagoons using these commands. I also updated the baseline documentation for changing the default lagoon to match the example given just for consistency.

Tagging in @AlannaBurke for visibility.